add pusher tests for all deep storages

+++ extensions/cloudfiles-extensions/src/main/java/io/druid/storage/cloudfiles/CloudFilesDataSegmentPusher.java
@@ -68,13 +68,17 @@ public class CloudFilesDataSegmentPusher implements DataSegmentPusher
   public DataSegment push(final File indexFilesDir, final DataSegment inSegment) throws IOException
   {
     final String segmentPath = CloudFilesUtils.buildCloudFilesPath(this.config.getBasePath(), inSegment);
-    final File descriptorFile = File.createTempFile("descriptor", ".json");
-    final File zipOutFile = File.createTempFile("druid", "index.zip");
-    final long indexSize = CompressionUtils.zip(indexFilesDir, zipOutFile);
 
-    log.info("Copying segment[%s] to CloudFiles at location[%s]", inSegment.getIdentifier(), segmentPath);
+    File descriptorFile = null;
+    File zipOutFile = null;
 
     try {
+      final File descFile = descriptorFile = File.createTempFile("descriptor", ".json");
+      final File outFile = zipOutFile = File.createTempFile("druid", "index.zip");
+
+      final long indexSize = CompressionUtils.zip(indexFilesDir, zipOutFile);
+
+      log.info("Copying segment[%s] to CloudFiles at location[%s]", inSegment.getIdentifier(), segmentPath);
       return CloudFilesUtils.retryCloudFilesOperation(
           new Callable<DataSegment>()
           {
@@ -82,17 +86,17 @@ public class CloudFilesDataSegmentPusher implements DataSegmentPusher
             public DataSegment call() throws Exception
             {
               CloudFilesObject segmentData = new CloudFilesObject(
-                  segmentPath, zipOutFile, objectApi.getRegion(),
+                  segmentPath, outFile, objectApi.getRegion(),
                   objectApi.getContainer()
               );
               log.info("Pushing %s.", segmentData.getPath());
               objectApi.put(segmentData);
 
-              try (FileOutputStream stream = new FileOutputStream(descriptorFile)) {
+              try (FileOutputStream stream = new FileOutputStream(descFile)) {
                 stream.write(jsonMapper.writeValueAsBytes(inSegment));
               }
               CloudFilesObject descriptorData = new CloudFilesObject(
-                  segmentPath, descriptorFile,
+                  segmentPath, descFile,
                   objectApi.getRegion(), objectApi.getContainer()
               );
               log.info("Pushing %s.", descriptorData.getPath());
@@ -123,11 +127,15 @@ public class CloudFilesDataSegmentPusher implements DataSegmentPusher
       throw Throwables.propagate(e);
     }
     finally {
-      log.info("Deleting zipped index File[%s]", zipOutFile);
-      zipOutFile.delete();
+      if (zipOutFile != null) {
+        log.info("Deleting zipped index File[%s]", zipOutFile);
+        zipOutFile.delete();
+      }
 
-      log.info("Deleting descriptor file[%s]", descriptorFile);
-      descriptorFile.delete();
+      if (descriptorFile != null) {
+        log.info("Deleting descriptor file[%s]", descriptorFile);
+        descriptorFile.delete();
+      }
     }
   }
 }
